Prepare negotiator for REST Interface
Add a test for functionality without bots. However, this test requires
changes to the code to work. The current version is not runable in
docker, because the Interface for the bots is missing. All important
locations and classes for the final implementation have been marked with
a "TODO: PA!", to indicate a change is required as soon as the interface
is available.